app/requirements.txt
asgiref==3.7.2
Django==4.2.3
djangorestframework==3.14.0
Pillow==10.0.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
pytz==2023.3
reportlab==4.0.4
six==1.16.0
sqlparse==0.4.4
gunicorn
qrbill
svglib
pdfrw
